removing hook
